Initial commit of OpenID Simple Registration extension module.
OpenID Simple Registation is an extension to the OpenID Authentication protocol that allows for very light-weight profile exchange. It is designed to pass eight commonly requested pieces of information when an End User goes to register a new account with a web service.

This module implements both, the Consumer and the Provider part of the specification.